+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/services/UIMetadataImportService.java
@@ -22,6 +22,7 @@ import ca.corefacility.bioinformatics.irida.model.sample.Sample;
 import ca.corefacility.bioinformatics.irida.model.sample.metadata.MetadataEntry;
 import ca.corefacility.bioinformatics.irida.ria.utilities.SampleMetadataStorage;
 import ca.corefacility.bioinformatics.irida.ria.utilities.SampleMetadataStorageRow;
+import ca.corefacility.bioinformatics.irida.ria.web.errors.SavedMetadataException;
 import ca.corefacility.bioinformatics.irida.ria.web.projects.ProjectControllerUtils;
 import ca.corefacility.bioinformatics.irida.service.ProjectService;
 import ca.corefacility.bioinformatics.irida.service.sample.MetadataTemplateService;
@@ -144,13 +145,25 @@ public class UIMetadataImportService {
 	 * @param projectId   {@link Long} identifier for the current project
 	 * @param sampleNames {@link List} of {@link String} sample names
 	 * @return {@link String} that returns a message and potential errors.
+	 * @throws SavedMetadataException if there is an error saving the metadata
 	 */
-	public SampleMetadataStorage saveProjectSampleMetadata(Locale locale, HttpSession session, Long projectId,
-			List<String> sampleNames) {
-		List<String> DEFAULT_HEADERS = ImmutableList.of("Sample Id", "ID", "Modified Date", "Modified On",
-				"Created Date", "Created On", "Coverage", "Project ID");
+	public String saveProjectSampleMetadata(Locale locale, HttpSession session, Long projectId,
+			List<String> sampleNames) throws SavedMetadataException {
+		List<String> DEFAULT_HEADERS = ImmutableList.of(
+				messageSource.getMessage("project.samples.table.sample-id", new Object[] {}, locale),
+				messageSource.getMessage("project.samples.table.id", new Object[] {}, locale),
+				messageSource.getMessage("project.samples.table.modified-date", new Object[] {}, locale),
+				messageSource.getMessage("project.samples.table.modified", new Object[] {}, locale),
+				messageSource.getMessage("project.samples.table.created-date", new Object[] {}, locale),
+				messageSource.getMessage("project.samples.table.created", new Object[] {}, locale),
+				messageSource.getMessage("project.samples.table.coverage", new Object[] {}, locale),
+				messageSource.getMessage("project.samples.table.project-id", new Object[] {}, locale));
 		Project project = projectService.read(projectId);
 		SampleMetadataStorage stored = (SampleMetadataStorage) session.getAttribute("pm-" + projectId);
+		boolean hasErrors = false;
+		String message;
+		int samplesUpdatedCount = 0;
+		int samplesCreatedCount = 0;
 
 		if (sampleNames != null) {
 			String sampleNameColumn = stored.getSampleNameColumn();
@@ -164,9 +177,11 @@ public class UIMetadataImportService {
 
 					if (row.getFoundSampleId() != null) {
 						sample = sampleService.getSampleBySampleName(project, name);
+						samplesUpdatedCount++;
 					} else {
 						sample = new Sample(name);
 						projectService.addSampleToProject(project, sample, true);
+						samplesCreatedCount++;
 					}
 
 					// Need to overwrite duplicate keys
@@ -193,11 +208,27 @@ public class UIMetadataImportService {
 					SampleMetadataStorageRow row = stored.getRow(sampleName, sampleNameColumn);
 					row.setError(e.getMessage());
 					row.setSaved(false);
+					hasErrors = true;
 				}
 			}
 		}
 
-		return stored;
+		if (hasErrors) {
+			throw new SavedMetadataException(stored);
+		}
+
+		message = ((samplesUpdatedCount == 1) ?
+				messageSource.getMessage("server.metadataimport.results.save.success.single-updated",
+						new Object[] { samplesUpdatedCount }, locale) :
+				messageSource.getMessage("server.metadataimport.results.save.success.multiple-updated",
+						new Object[] { samplesUpdatedCount }, locale));
+		message += (samplesCreatedCount == 1) ?
+				messageSource.getMessage("server.metadataimport.results.save.success.single-created",
+						new Object[] { samplesCreatedCount }, locale) :
+				messageSource.getMessage("server.metadataimport.results.save.success.multiple-created",
+						new Object[] { samplesCreatedCount }, locale);
+
+		return message;
 	}
 
 	/**